Converted backend to ts, moved into folders
