+++ backend/package.json
@@ -2,6 +2,7 @@
     "name": "e-immigrate2",
     "version": "0.1.0",
     "private": true,
+    "type": "module",
     "dependencies": {
         "@sendgrid/mail": "^8.1.1",
         "axios": "^1.6.8",
@@ -14,10 +15,30 @@
         "jsonwebtoken": "^9.0.2",
         "mongodb": "^6.5.0",
         "mongoose": "^8.2.2",
-        "read-excel-file": "^5.7.1"
+        "node-fetch": "^3.3.2",
+        "read-excel-file": "^5.7.1",
+        "zod": "^3.22.4"
+    },
+    "devDependencies": {
+        "@types/bcrypt": "^5.0.2",
+        "@types/cors": "^2.8.17",
+        "@types/express": "^4.17.21",
+        "@types/express-fileupload": "^1.4.4",
+        "@types/jsonwebtoken": "^9.0.6",
+        "@types/node": "^20.11.28",
+        "nodemon": "^3.1.0",
+        "typescript": "^5.4.2"
     },
     "scripts": {
-        "start": "node ./index.js"
+        "compile": "tsc",
+        "start": "node ./dist/index.js",
+        "start:watch": "nodemon --delay 500ms -L --exec npm run start",
+        "build": "npm run compile",
+        "build-and-serve": "npm run build && npm run start",
+        "build-and-serve:watch": "npm run build && npm run start:watch",
+        "upload-from-excel:questions": "node ./dist/uploadQuestionnaireFromExcel.js",
+        "upload-from-excel:translated": "node ./dist/uploadTranslatedContentFromExcel.js",
+        "prepare-from-excel": "npm run upload-from-excel:questions && npm run upload-from-excel:translated"
     },
     "eslintConfig": {
         "extends": "react-app"
