+++ backend/src/LanguageOptions.ts
@@ -1,0 +1,89 @@
+import { EmailContentLanguages } from "./routes/sendEmail/emailContent.js";
+
+export const WorkshopTitle = 'CIIT_Workshop_Spring_2022';
+
+export type LanguageOptionCodes = EmailContentLanguages | 'ko' | 'pa';
+type LanguageOption = {
+    code: LanguageOptionCodes;
+    full: string;
+    englishName: string;
+};
+export const LanguageOptions: Array<LanguageOption> = [
+    {
+        code: 'en',
+        full: 'English',
+        englishName: 'English',
+    },
+    {
+        code: 'es',
+        full: 'Español',
+        englishName: 'Spanish',
+    },
+    {
+        code: 'vi',
+        full: 'Tiếng Việt',
+        englishName: 'Vietnamese',
+    },
+    {
+        code: 'tl',
+        full: 'Wikang Tagalog',
+        englishName: 'Tagalog',
+    },
+    {
+        code: 'cmn',
+        full: '普通话',
+        englishName: 'Mandarin',
+    },
+    {
+        code: 'yue',
+        full: '廣東話',
+        englishName: 'Cantonese',
+    },
+    {
+        code: 'ru',
+        full: 'русский',
+        englishName: 'Russian',
+    },
+    {
+        code: 'am',
+        full: 'አማርኛ',
+        englishName: 'Amharic',
+    },
+    {
+        code: 'ar',
+        full: 'العربية',
+        englishName: 'Arabic',
+    },
+
+    {
+        code: 'fa',
+        full: 'فارسی',
+        englishName: 'Farsi',
+    },
+    {
+        code: 'hi',
+        full: 'हिंदी',
+        englishName: 'Hindi',
+    },
+    {
+        code: 'ko',
+        full: '한국어',
+        englishName: 'Korean',
+    },
+
+    {
+        code: 'ps',
+        full: 'پښتو',
+        englishName: 'Pashto',
+    },
+    {
+        code: 'pa',
+        full: 'ਪੰਜਾਬੀ',
+        englishName: 'Punjabi',
+    },
+    {
+        code: 'pt',
+        full: 'português',
+        englishName: 'Portuguese',
+    },
+];
