+++ backend/src/custom-typings/excel4node.d.ts
@@ -1,0 +1,3 @@
+export default {
+    Workbook: unknown,
+}
