+++ backend/src/features/emails/emailSender.ts
@@ -1,0 +1,6 @@
+export function getRequiredEmailSender() {
+    const output = process.env.SENDER_EMAIL;
+    if (output == null || output === '') throw new Error(`Missing the 'SENDER_EMAIL' environment variable`);
+    return output;
+}
+export const emailSender = getRequiredEmailSender();
